Stopped sync journals to sites from always deleting sites objects.
Added --delete option for when this is required. Closes #28.

+++ src/utils/management/commands/sync_journals_to_sites.py
@@ -11,6 +11,14 @@ class Command(BaseCommand):
 
     help = "Synchronizes configured journal and press URLs with sites."
 
+    def add_arguments(self, parser):
+        """ Adds arguments to Django's management command-line parser.
+
+        :param parser: the parser to which the required arguments will be added
+        :return: None
+        """
+        parser.add_argument('--delete', action='store_true', default=False)
+
     def handle(self, *args, **options):
         """Synchronizes journals and presses with sites.
 
@@ -20,20 +28,19 @@ class Command(BaseCommand):
         """
 
         translation.activate('en')
-        site_models.Site.objects.all().delete()
+        if options.get('delete'):
+            site_models.Site.objects.all().delete()
 
         journals = journal_models.Journal.objects.all()
 
         print("Journals:")
         for journal in journals:
             print(journal.name, journal.domain)
-            new_site = site_models.Site(name=journal.code, domain=journal.domain)
-            new_site.save()
+            new_site, c = site_models.Site.objects.get_or_create(name=journal.code, domain=journal.domain)
             print('Journal with ID {0} [{1}]: {2}. SYNCED.'.format(journal.id, journal.name, journal.domain))
 
         print("\nPress:")
         p = press_models.Press.get_press(None)
 
-        new_site = site_models.Site(name=p.name, domain=p.domain)
-        new_site.save()
+        new_site, c = site_models.Site.objects.get_or_create(name=p.name, domain=p.domain)
         print("Press [{0}]: {1}. SYNCED.".format(p.name, p.domain))
